+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 * [BREAKING] Renamed `Action::Configuration` + `Action.config` -> `Axn::Configuration` + `Axn.config`
 * [BREAKING] Move `Axn::Util` to `Axn::Internal::Logging`
 * [BREAKING] !! Move all `Action` to `Axn` (notably `include Action` is now `include Axn`)
-* [FEAT] Continues to support plain ruby usage, but when used alongside Rails now includes a Rails Engine integrate automaticaly (e.g. providing generators).
-* [FEAT] Added Rails generator `rails generate axn Some::Action::Name foo bar` to create action classes with expectations
+* [FEAT] Continues to support plain ruby usage, but when used alongside Rails now includes a Rails Engine integrate automatically (e.g. providing generators).
+  * Added Rails generator `rails generate axn Some::Action::Name foo bar` to create action classes with expectations
+  * Autoload actions from `app/actions` (add config.rails.app_actions_autoload_namespace to allow setting custom namespace)
 * [INTERNAL] Clearer hooks for supporting additional background providers in the future
 * [BREAKING] spec_helpers: removed rarely used `build_axn`; renamed existing `build_action` -> `build_axn`